--- a/hw10_program_optimization/go.mod
+++ b/hw10_program_optimization/go.mod
@@ -2,10 +2,15 @@ module github.com/inenagl/hw-Go-Prof/hw10_program_optimization
 
 go 1.19
 
-require github.com/stretchr/testify v1.7.0
+require (
+	github.com/stretchr/testify v1.7.0
+	github.com/tidwall/gjson v1.16.0
+)
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/tidwall/match v1.1.1 // indirect
+	github.com/tidwall/pretty v1.2.1 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )